change product weight field type from Integer to Long

+++ src/main/java/com/journal/florist/backend/feature/product/dto/product/AddProductRequest.java
@@ -16,7 +16,7 @@ public class AddProductRequest {
     private String description;
     private Integer stock;
     private BigDecimal price;
-    private Integer weight;
+    private Long weight;
     private String material;
     private BigDecimal materialPrice;
 }
